Refactor project to professional standards
This commit implements a comprehensive set of improvements to elevate the project to an enterprise-ready standard.

Key changes include:
- Repository Structure: Consolidated the test suite into a single top-level `tests` directory, removed temporary and generated files, and added a LICENSE.
- Documentation: Replaced the README with a professional and functional version containing setup and usage instructions.
- Code Quality & Configuration:
  - Implemented secret management using a `.env` file and `python-dotenv`.
  - Replaced all `print()` statements in the core application logic with a structured `rich` logger.
  - Integrated `Flask-Migrate` for robust database schema migrations and removed the old `init_db.py` script.
- Bug Fixes:
  - Refactored the Flask application to use the `create_app` factory pattern, fixing a critical startup bug related to extension initialization.
  - Corrected the path to the dependency installation script in the CLI entry point.
