+++ .env.example
@@ -1,0 +1,2 @@
+SECRET_KEY='generate_a_strong_random_key_for_production'
+DATABASE_URL='sqlite:///cyberhunter.db'
