+++ config.py
@@ -1,7 +1,25 @@
 import os
+from dotenv import load_dotenv
+
+# Load variables from .env file
+load_dotenv()
 
 class Config:
-    SECRET_KEY = os.environ.get('SECRET_KEY') or 'a-very-secret-key-that-should-be-changed'
-    DB_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'cyberhunter.db')
-    SQLALCHEMY_DATABASE_URI = f'sqlite:///{DB_PATH}'
+    # It's crucial that you set this in your environment or a .env file
+    SECRET_KEY = os.environ.get('SECRET_KEY')
+    if not SECRET_KEY:
+        raise ValueError("No SECRET_KEY set for Flask application. Please set it in your environment or a .env file.")
+
+    # Use DATABASE_URL from .env or default to a local sqlite db
+    DATABASE_URL = os.environ.get('DATABASE_URL') or 'sqlite:///cyberhunter.db'
+
+    SQLALCHEMY_DATABASE_URI = DATABASE_URL
     SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+    # Configuration for Flask-Mail
+    MAIL_SERVER = os.environ.get('MAIL_SERVER')
+    MAIL_PORT = int(os.environ.get('MAIL_PORT', 587))
+    MAIL_USE_TLS = os.environ.get('MAIL_USE_TLS', 'true').lower() in ['true', 'on', '1']
+    MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
+    MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
+    MAIL_DEFAULT_SENDER = os.environ.get('MAIL_DEFAULT_SENDER')
