--- final_recon_data.json
@@ -1,10 +1,0 @@
-{
-    "domain": "example.com",
-    "master_subdomains": [],
-    "live_hosts": [],
-    "screenshots": [],
-    "technology_and_ports": {},
-    "js_findings": [],
-    "cloud_assets": [],
-    "github_findings": []
-}
