+++ requirements.txt
@@ -26,3 +26,5 @@ python-docx
 APScheduler
 Flask-Mail
 rich
+python-dotenv
+Flask-Migrate
