+++ tests/e2e/test_graph_view.py
