+++ tests/test_cloud_assets.py
