+++ tests/test_scope_validator.py
