--- web_server.log
@@ -1,10 +1,0 @@
-User table not found. Creating all tables...
-Database tables created.
-Creating default user 'test' with password 'password'
-Traceback (most recent call last):
-  File "/app/run_web.py", line 397, in <module>
-    init_database()
-  File "/app/run_web.py", line 45, in init_database
-    password_hash = app.bcrypt.generate_password_hash('password').decode('utf-8')
-                    ^^^^^^^^^^
-AttributeError: 'Flask' object has no attribute 'bcrypt'
